refactor: one shape for every payload
Every wire serial is a u32 and every key tag a u16; counts are u64 named
added/deleted/unchanged, and only a diff says removed; a zone delete
reports records_deleted and versions_deleted; a record lives at
/records/{id}. A field naming another entity is <entity>_name
(token_name, tsig_key_name, notify_key_name, policy_name,
catalog_zone_name). A response's Option is emitted as null, never skipped.
The NOTIFY check reports address and error, the DNSSEC status travels
bare, and status-like fields are enums with a schema (SecondaryStatus,
RecordChange, DsState, and core's key role, key state, denial, and change
source). The zone and record listings derive their query parameters from
the filter structs. serial_to_i32 joins serial_to_u32 in core, and the
grant path parameter is NameIdParam. CLAUDE.md records the conventions;
the OpenAPI document is regenerated.

+++ crates/bindizr-service/src/record/delete.rs
@@ -260,7 +260,7 @@ impl RecordService {
             let response = DeleteRecordsResponse {
                 applied: !filter.dry_run,
                 dry_run: filter.dry_run,
-                deleted: matched.len(),
+                deleted: matched.len() as u64,
                 records: matched
                     .iter()
                     .map(|record| GetRecordResponse::from_record_and_zone_name(record, &zone.name))
